feat: modularize codebase into smaller files
Splits the monolithic [`src/main.rs`](/src/main.rs) into several smaller
files to aid in comprehension.
